requirements.txt
streamlit==1.32.0
tensorflow-cpu==2.16.0
numpy==1.26.4
Pillow==10.4.0